requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
aiokafka==0.10.0
numpy==1.26.3
aiohttp==3.9.1
